Altered some Xcode 4.4 warnings, SR uses a depreciated method of some sort in 10.8 and with the development of MASShortcut I don't feel like fixing it. Also added a doc html file I found somewhere

+++ ShortcutRecorderTest/KSAppDelegate.m
@@ -59,7 +59,7 @@
     hotKeyCenter = [PTHotKeyCenter sharedCenter];
     
     signed short code = newKeyCombo.code;
-    unsigned int flags = [aRecorder cocoaToCarbonFlags:newKeyCombo.flags];
+    NSUInteger flags = [aRecorder cocoaToCarbonFlags:newKeyCombo.flags];
     akeyCombo = [[PTKeyCombo alloc] initWithKeyCode:code modifiers:flags];
     
     if (aRecorder == shortcutRecorder) {
